refactor code to reuse datacenter and datastore fetch functions
also fix logging
