+++ pkg/check/util.go
@@ -1,0 +1,48 @@
+package check
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/vmware/govmomi/find"
+	"github.com/vmware/govmomi/object"
+	"github.com/vmware/govmomi/property"
+	"github.com/vmware/govmomi/vim25/mo"
+	vim "github.com/vmware/govmomi/vim25/types"
+)
+
+func getDatacenter(ctx *CheckContext, dcName string) (*object.Datacenter, error) {
+	tctx, cancel := context.WithTimeout(ctx.Context, *Timeout)
+	defer cancel()
+	finder := find.NewFinder(ctx.VMClient, false)
+	dc, err := finder.Datacenter(tctx, dcName)
+	if err != nil {
+		return nil, fmt.Errorf("failed to access datacenter %s: %s", dcName, err)
+	}
+	return dc, nil
+}
+
+func getDataStoreByName(ctx *CheckContext, dsName string, dc *object.Datacenter) (*object.Datastore, error) {
+	tctx, cancel := context.WithTimeout(ctx.Context, *Timeout)
+	defer cancel()
+	finder := find.NewFinder(ctx.VMClient, false)
+	finder.SetDatacenter(dc)
+	ds, err := finder.Datastore(tctx, dsName)
+	if err != nil {
+		return nil, fmt.Errorf("failed to access datastore %s: %s", dsName, err)
+	}
+	return ds, nil
+}
+
+func getDatastore(ctx *CheckContext, ref vim.ManagedObjectReference) (mo.Datastore, error) {
+	var dsMo mo.Datastore
+	pc := property.DefaultCollector(ctx.VMClient)
+	properties := []string{DatastoreInfoProperty, SummaryProperty}
+	tctx, cancel := context.WithTimeout(ctx.Context, *Timeout)
+	defer cancel()
+	err := pc.RetrieveOne(tctx, ref, properties, &dsMo)
+	if err != nil {
+		return dsMo, err
+	}
+	return dsMo, nil
+}
